Resolved Conflicts:
	admin/gitlab-ci/api-client.matrix/gromacs-master.gitlab-ci.yml
	cmake/gmxVersionInfo.cmake
	docs/release-notes/index.rst
	python_packaging/src/gmxapi/simulation/mdrun.py

+++ api/legacy/include/gromacs/math/functions.h
@@ -2,7 +2,7 @@
  * This file is part of the GROMACS molecular simulation package.
  *
  * Copyright (c) 2015,2016,2018,2019,2020 by the GROMACS development team.
- * Copyright (c) 2021, by the GROMACS development team, led by
+ * Copyright (c) 2021,2022, by the GROMACS development team, led by
  * Mark Abraham, David van der Spoel, Berk Hess, and Erik Lindahl,
  * and including many others, as listed in the AUTHORS file in the
  * top-level source directory and at http://www.gromacs.org.
@@ -465,12 +465,12 @@ static inline int roundToInt(double x)
 //! Round float to int64_t
 static inline int64_t roundToInt64(float x)
 {
-    return static_cast<int>(rintf(x));
+    return static_cast<int64_t>(rintf(x));
 }
 //! Round double to int64_t
 static inline int64_t roundToInt64(double x)
 {
-    return static_cast<int>(rint(x));
+    return static_cast<int64_t>(rint(x));
 }
 
 //! \brief Check whether \p v is an integer power of 2.
